moved calendar event and exercise event
